+++ StripeCore/StripeCore/Source/Connections Bindings/FinancialConnectionsConsumer.swift
@@ -13,19 +13,22 @@ import Foundation
     @_spi(STP) public let emailAddress: String
     @_spi(STP) public let redactedFormattedPhoneNumber: String
     @_spi(STP) public let verificationSessions: [VerificationSession]
+    @_spi(STP) public let linkBrand: LinkBrand?
 
     @_spi(STP) public init(
         publishableKey: String?,
         clientSecret: String,
         emailAddress: String,
         redactedFormattedPhoneNumber: String,
-        verificationSessions: [VerificationSession]
+        verificationSessions: [VerificationSession],
+        linkBrand: LinkBrand? = nil
     ) {
         self.publishableKey = publishableKey
         self.clientSecret = clientSecret
         self.emailAddress = emailAddress
         self.redactedFormattedPhoneNumber = redactedFormattedPhoneNumber
         self.verificationSessions = verificationSessions
+        self.linkBrand = linkBrand
     }
 }