+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/PaymentSheet+APIMockTest.swift
@@ -76,13 +76,13 @@ final class PaymentSheetAPIMockTest: APIStubbedTestCase {
         static var linkPaymentOption: PaymentSheet.PaymentOption {
             let exampleBillingEmail = "test@example.com"
             return .link(option: .withPaymentDetails(
-                brand: .link,
-                account: .init(
-                    email: exampleBillingEmail,
-                    session: .init(
-                        clientSecret: "cs_xxx",
-                        emailAddress: exampleBillingEmail,
-                        redactedFormattedPhoneNumber: "(***) *** **55",
+                    brand: .link,
+                    account: .init(
+                        email: exampleBillingEmail,
+                        session: .make(
+                            clientSecret: "cs_xxx",
+                            emailAddress: exampleBillingEmail,
+                            redactedFormattedPhoneNumber: "(***) *** **55",
                         unredactedPhoneNumber: "(555) 555-5555",
                         phoneNumberCountry: "US",
                         verificationSessions: [.init(type: .sms, state: .verified)],
@@ -188,7 +188,7 @@ final class PaymentSheetAPIMockTest: APIStubbedTestCase {
                     brand: .link,
                     account: .init(
                         email: "test@example.com",
-                        session: .init(
+                        session: .make(
                             clientSecret: "cs_xxx",
                             emailAddress: "test@example.com",
                             redactedFormattedPhoneNumber: "(***) *** **55",
@@ -317,7 +317,7 @@ final class PaymentSheetAPIMockTest: APIStubbedTestCase {
                 brand: .link,
                 account: .init(
                     email: "test@example.com",
-                    session: .init(
+                    session: .make(
                         clientSecret: "cs_xxx",
                         emailAddress: "test@example.com",
                         redactedFormattedPhoneNumber: "(***) *** **55",