+++ StripePayments/StripePayments/Source/Internal/Helpers/ConnectionsSDKAvailability.swift
@@ -90,8 +90,6 @@ import UIKit
 }
 
 final class StubbedConnectionsSDKInterface: FinancialConnectionsSDKInterface {
-    required init() {}
-
     func presentFinancialConnectionsSheet(
         apiClient: STPAPIClient,
         clientSecret: String,
@@ -99,7 +97,7 @@ final class StubbedConnectionsSDKInterface: FinancialConnectionsSDKInterface {
         existingConsumer: FinancialConnectionsConsumer?,
         style: FinancialConnectionsStyle,
         elementsSessionContext: ElementsSessionContext?,
-        brand: LinkBrand?,
+        linkBrand: LinkBrand?,
         onEvent: ((FinancialConnectionsEvent) -> Void)?,
         from presentingViewController: UIViewController,
         completion: @escaping (FinancialConnectionsSDKResult) -> Void