+++ CHANGELOG.md
@@ -2,6 +2,14 @@ The next release's version bump will so far be:
 PATCH
 
 ## X.Y.Z - changes pending release
+### PaymentSheet
+* [Fixed] Fixed Japan address form missing the city field. ([#6506](https://github.com/stripe/stripe-ios/issues/6506))
+
+### Address Element
+* [Fixed] Fixed Japan address form missing the city field. ([#6506](https://github.com/stripe/stripe-ios/issues/6506))
+
+## 25.15.0 2026-05-18
+* Added support for [Onelink](https://support.stripe.com/questions/what-is-onelink).
 
 ## 25.14.0 2026-05-11
 ### PaymentSheet